+++ packages/atomic-router/src/index.ts
@@ -1,2 +1,3 @@
 export { freshChain } from './fresh';
 export { startChain } from './start';
+export { barrierChain } from './barrier';