+++ packages/core/src/cache/key/key.ts
@@ -50,7 +50,7 @@ export function queryUniqId(query: Query<any, any, any>) {
   }
 
   throw new Error(
-    'Query does not have sid or uniq name, which is required for caching, read more https://farfetched.pages.dev/recipes/sids.html'
+    'Query does not have sid or uniq name, which is required for caching, read more https://ff.effector.dev/recipes/sids.html'
   );
 }