+++ packages/atomic-router/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @farfetched/atomic-router
 
+## 0.12.4
+
+### Patch Changes
+
+- 42ce0f0: Add missed export of `barrierChain`
+  - @farfetched/core@0.12.4
+
+## 0.12.3
+
+### Patch Changes
+
+- 16f68b2: Update formatting (bump prettier to v3)
+- 1f59bb3: Fix race in `barrierChain`
+- Updated dependencies [16f68b2]
+- Updated dependencies [d0fc25b]
+- Updated dependencies [1f59bb3]
+  - @farfetched/core@0.12.3
+
 ## 0.12.2
 
 ### Patch Changes